LoopBuffer: Delete old files
