+++ src/main/scala/xiangshan/frontend/IFU.scala
@@ -48,7 +48,7 @@ class IFU extends XSModule with HasIFUConst
   val io = IO(new IFUIO)
   val bpu = BPU(EnableBPU)
   val pd = Module(new PreDecode)
-  val loopBuffer = if(EnableLB) { Module(new NewLoopBuffer) } else { Module(new FakeLoopBuffer) }
+  val loopBuffer = if(EnableLB) { Module(new LoopBuffer) } else { Module(new FakeLoopBuffer) }
 
   val if2_redirect, if3_redirect, if4_redirect = WireInit(false.B)
   val if1_flush, if2_flush, if3_flush, if4_flush = WireInit(false.B)
